Feat(Elevator) Added Elevator Prefab
Added Elevator, as well as several new prefabs
