+++ Assets/3D Models/Materials/ElevatorTexture.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: df4c250c77cf365439801e15465cc2e4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
