+++ Assets/Prefabs/Elevator.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1a8fc109064b8e64e84c9139fcf4884a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
