+++ Assets/Prefabs/MetalDoor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1e3deaed35459c14cbec11c9d3063579
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
