+++ Assets/Prefabs/WoodDoor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3817914c204930f4a85d0f571f897ac6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
