+++ Assets/Scripts/Elevator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9528b229e73a56b45b80164d7c2a7143
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
